+++ Source/Headers/MDCButton.h
@@ -1,0 +1,4 @@
+#import <UIKit/UIKit.h>
+
+@interface MDCButton : UIButton
+@end
