+++ Source/Headers/YTAssetLoader.h
@@ -1,0 +1,6 @@
+#import <UIKit/UIKit.h>
+
+@interface YTAssetLoader : NSObject
+- (instancetype)initWithBundle:(NSBundle *)bundle;
+- (UIImage *)imageNamed:(NSString *)image;
+@end
