+++ Source/Headers/YTCommonColorPalette.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTCommonColorPalette : NSObject
+- (UIColor *)textPrimary;
+@end
