+++ Source/Headers/YTIAccessibilityData.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIAccessibilityData : NSObject
+@property (nonatomic, copy, readwrite) NSString *label;
+@end
