+++ Source/Headers/YTIAccessibilitySupportedDatas.h
@@ -1,0 +1,5 @@
+#import "YTIAccessibilityData.h"
+
+@interface YTIAccessibilitySupportedDatas : NSObject
+@property (nonatomic, strong, readwrite) YTIAccessibilityData *accessibilityData;
+@end
