+++ Source/Headers/YTIBrowseEndpoint.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIBrowseEndpoint : NSObject
+@property (nonatomic, copy, readwrite) NSString *browseId;
+@end
