+++ Source/Headers/YTICommand.h
@@ -1,0 +1,5 @@
+#import "YTIBrowseEndpoint.h"
+
+@interface YTICommand : NSObject
+@property (nonatomic, readwrite, strong) YTIBrowseEndpoint *browseEndpoint;
+@end
