+++ Source/Headers/YTIFormatStream.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIFormatStream : NSObject
+@property (nonatomic, copy, readwrite) NSString *URL;
+@end
