+++ Source/Headers/YTIFormattedString.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIFormattedString : NSObject
++ (id)formattedStringWithString:(id)arg1;
+@end
