+++ Source/Headers/YTIPivotBarRenderer.h
@@ -1,0 +1,5 @@
+#import "YTIPivotBarSupportedRenderers.h"
+
+@interface YTIPivotBarRenderer : NSObject
+- (NSMutableArray <YTIPivotBarSupportedRenderers *> *)itemsArray;
+@end
