+++ Source/Headers/YTIPivotBarSupportedRenderers.h
@@ -1,0 +1,5 @@
+#import "YTIPivotBarItemRenderer.h"
+
+@interface YTIPivotBarSupportedRenderers : NSObject
+@property (nonatomic, retain) YTIPivotBarItemRenderer *pivotBarItemRenderer;
+@end
