+++ Source/Headers/YTIPlayerOverlayActionSupportedRenderers.h
@@ -1,0 +1,5 @@
+#import "YTIButtonRenderer.h"
+
+@interface YTIPlayerOverlayActionSupportedRenderers : NSObject
+@property (nonatomic, strong, readwrite) YTIButtonRenderer *buttonRenderer;
+@end
