+++ Source/Headers/YTIPlayerOverlayRenderer.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIPlayerOverlayRenderer : NSObject
+@property (nonatomic, strong, readwrite) NSMutableArray *actionsArray;
+@end
