+++ Source/Headers/YTIStreamingData.h
@@ -1,0 +1,6 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIStreamingData : NSObject
+@property (nonatomic, copy, readwrite) NSString *hlsManifestURL;
+@property (nonatomic, copy, readwrite) NSMutableArray *adaptiveFormatsArray;
+@end
