+++ Source/Headers/YTIThumbnailDetails.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIThumbnailDetails : NSObject
+@property (nonatomic, strong, readwrite) NSMutableArray *thumbnailsArray;
+@end
