+++ Source/Headers/YTIThumbnailDetails_Thumbnail.h
@@ -1,0 +1,6 @@
+#import <Foundation/Foundation.h>
+
+@interface YTIThumbnailDetails_Thumbnail : NSObject
+@property (nonatomic, copy, readwrite) NSString *URL;
+@property (nonatomic, assign, readwrite) unsigned int width;
+@end
