+++ Source/Headers/YTMActionSheetController.h
@@ -1,0 +1,5 @@
+#import "YTActionSheetController.h"
+
+@interface YTMActionSheetController : YTActionSheetController
++ (instancetype)musicActionSheetController;
+@end
