+++ Source/Headers/YTMBrowseViewController.h
@@ -1,0 +1,4 @@
+#import <UIKit/UIKit.h>
+
+@interface YTMBrowseViewController : UIViewController
+@end
