+++ Source/Headers/YTMNowPlayingView.h
@@ -1,0 +1,5 @@
+#import "YTMPlayerControlsView.h"
+
+@interface YTMNowPlayingView : UIView
+@property (nonatomic, assign, readonly) YTMPlayerControlsView *playerControlsView;
+@end
