+++ Source/Headers/YTMPlayerControlsView.h
@@ -1,0 +1,8 @@
+#import "MDCButton.h"
+
+@interface YTMPlayerControlsView : UIView
+@property (nonatomic, assign) MDCButton *prevButton;
+@property (nonatomic, assign) MDCButton *nextButton;
+@property (nonatomic, assign) MDCButton *seekForwardButton;
+@property (nonatomic, assign) MDCButton *seekBackwardButton;
+@end
