+++ Source/Headers/YTMToastController.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTMToastController : NSObject
+- (void)showMessage:(NSString *)message;
+@end
