+++ Source/Headers/YTMWatchViewController.h
@@ -1,0 +1,5 @@
+#import "YTPlayerViewController.h"
+
+@interface YTMWatchViewController : UIViewController
+@property (nonatomic, weak, readwrite) YTPlayerViewController *playerViewController;
+@end
