+++ Source/Headers/YTPivotBarItemView.h
@@ -1,0 +1,5 @@
+#import "YTQTMButton.h"
+
+@interface YTPivotBarItemView : UIView
+@property (nonatomic, strong, readwrite) YTQTMButton *navigationButton;
+@end
