+++ Source/Headers/YTPivotBarView.h
@@ -1,0 +1,6 @@
+#import <UIKit/UIKit.h>
+#import "YTIPivotBarRenderer.h"
+
+@interface YTPivotBarView : UIView
+@property (nonatomic, strong, readwrite) YTIPivotBarRenderer *renderer;
+@end
