+++ Source/Headers/YTPivotBarViewController.h
@@ -1,0 +1,5 @@
+#import <UIKit/UIKit.h>
+
+@interface YTPivotBarViewController : UIViewController
+- (void)selectItemWithPivotIdentifier:(NSString *)pivotIndentifier;
+@end
