+++ Source/Headers/YTPlayabilityResolutionUserActionUIController.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTPlayabilityResolutionUserActionUIController : NSObject
+- (void)confirmAlertDidPressConfirm;
+@end
