+++ Source/Headers/YTPlayerResponse.h
@@ -1,0 +1,5 @@
+#import "YTIPlayerResponse.h"
+
+@interface YTPlayerResponse : NSObject
+@property (nonatomic, assign, readonly) YTIPlayerResponse *playerData;
+@end
