+++ Source/Headers/YTPlayerViewController.h
@@ -1,0 +1,8 @@
+#import <UIKit/UIKit.h>
+#import "YTPlayerResponse.h"
+
+@interface YTPlayerViewController : UIViewController
+@property (nonatomic, assign, readonly) YTPlayerResponse *playerResponse;
+@property (readonly, nonatomic) NSString *contentVideoID;
+@property (nonatomic, assign, readonly) CGFloat currentVideoTotalMediaTime;
+@end
