+++ Source/Headers/YTQTMButton.h
@@ -1,0 +1,5 @@
+#import <UIKit/UIKit.h>
+
+@interface YTQTMButton : UIButton
+- (void)setSizeWithPaddingAndInsets:(BOOL)sizeWithPaddingAndInsets;
+@end
