+++ Source/Headers/YTUIResources.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+
+@interface YTUIResources : NSObject
++ (UIImage *)youtubePremiumLogoWithPageStyle:(NSInteger)pageStyle;
+@end
