+++ Source/Prefs/PlayerSettingsController.h
@@ -1,5 +1,5 @@
-#import "UIKit/UIKit.h"
-#import "Localization.h"
+#import <UIKit/UIKit.h>
+#import "../Headers/Localization.h"
 
 @interface PlayerSettingsController : UIViewController <UITableViewDelegate, UITableViewDataSource, UITextFieldDelegate>
 @property (nonatomic, strong) UITableView* tableView;
