+++ Source/Prefs/PremiumSettingsController.h
@@ -1,5 +1,5 @@
-#import "UIKit/UIKit.h"
-#import "Localization.h"
+#import <UIKit/UIKit.h>
+#import "../Headers/Localization.h"
 
 @interface PremiumSettingsController : UIViewController <UITableViewDelegate, UITableViewDataSource>
 @property (nonatomic, strong) UITableView* tableView;
